creating checkouts_controller per Stripe recommendation; beginning to transition from Stripe-hosted checkout page to my own checkout page with Stripe elements-- created that checkout page in checkouts views dir; routing 'Check Out' button to this page instead of requesting the Stripe-hosted page
