+++ app/helpers/checkouts_helper.rb
@@ -1,0 +1,2 @@
+module CheckoutsHelper
+end
