+++ app/views/checkouts/create.html.erb
@@ -1,0 +1,4 @@
+<div>
+  <h1 class="font-bold text-4xl">Checkouts#create</h1>
+  <p>Find me in app/views/checkouts/create.html.erb</p>
+</div>
