+++ app/views/checkouts/new.html.erb
@@ -1,0 +1,1 @@
+<p>Here is where you'll embed your Stripe elements!</p>
